Huge plethora of fixes to Less Lethal equipment.
- Fixed bug with Less Lethal Shotgun, Melee, Triple Baton, and direct
grenade hits causing an "Unauthorized Use of Force" or "Unauthorized Use
of Deadly Force" when the target is considered a threat. (This was due to
the character being stunned BEFORE the damage was taking place)

- CSBallBase moved to Projectiles, away from Ammo (CSBallBase is the
projectile shot from a Pepperball gun, not the ammo that feeds into the
pepperball gun)

- Deploy Grenade Launcher now works correctly for the ARWEN 37

- Deploy Taser now works correctly if the AI has it equipped as a primary
weapon.

- Deploy Less Lethal Shotgun now works correctly if the AI has it equipped
as a secondary weapon.

- Less Lethal Shotgun now derives from BeanbagShotgunBase; instead of
migrating LL properties to ammo, it will remain part of the weapon, but be
relegated to special "Less Lethal" versions of weapons. The intention is
that the player will be able to pick Beanbags as part of the ammo list,
which changes to the green Less Lethal version. Navigating away from the
ammo or weapon would cycle as normal.
Or maybe that's too complicated and we should keep the Less Lethal
versions in the rotation? Either way, we can shorten the list later.