Added the base code for the damage system
How it is right now:
Ammo types get a certain Armor penetration level
Armor and helmets get a certain Armor protection level
If Armor penetration is bigger or equal to Armor protection, the armor/helmet is penetrated. Otherwise, the armor/helmet stops the bullet
The new damage system only applies to select weapons (Right now, only the M4 and M1911 get this system)
Experiment a bit with the diferent ammo types and armors to see how the weapons behave

Damage formula as of now:
1 - (X/B)
X is the body part factor, 1 for head. For chest it is: 100 for buckshot, 125 for JHP, 150 for JSP, 175 for FMJ and 200 for AP. 1350 for arms. 900 for leg.
B is the bullet's kill energy. It already calculates the velocity of the bullet based on the distance.

If you need a longer explaination ask me.