Made the High skill enemy AI badass and other AI changes
Lower chance of screaming due to a nearby hit when complied
Higher chances of reaction to a thrown grenade
Higher Accuracy
Chances of firing for extended periods of time, simulating suppression fire
Sent the chances of flinching due to gunshots to the ground.

To SWAT
Fixed a typo that always made the AI take cover, even with point blank targets

To all AIs
Distance to Stop Moving is now 0. Might help to reduce the SWAT won't shoot pointblank targets
Suspects might suddenly disengage SWAT after firing
Reduced the time period before suspects open up on hostages to force the player to not suck and be fast on room entries
Increased the time suspects peek out of cover
Reduced the delay between each burst of fire
Added a 1-2 second delay to the AI getting pepper sprayed