Accuracy System Overhaul
All weapons have an overhaul. In short:
The weapon's Walk-to-Run penalties has been increased
The weapon's Running minimum accuracy was increased (Aim Error is reduced)
The weapon's recovery rates have been increased (faster recoveries)

The point is to make the player be able to move at full speed and be able to clear rooms like the SWAT AI (as well as real SWAT Officers) does.