CS gas crash possible fix
i've found that effect duration code is probably passing negative or zero duration to time , causing effect to crash. 

https://github.com/eezstreet/SWATEliteForce/blob/52ff0e3158ff1842a5ea36e65e73ecfe79bd9aba/Source/Game/SwatGame/Classes/SwatPlayer.uc#L2535

DistanceEffect = (600 - Distance)/(600);

I've so setted the radius to 599 to keep the distance from player to the projectile always positive. 

on the contrary , setting the radius  over 600 makes constant crash if player is more distant than 600.