Improved Lightstick
This is an attempt to improve the accuracy of Lightsticks.
In SWAT 4, lighsticks lack the brightness that they have IRL.
Also, the duration of lightsticks in the game is too short compared to the real ones. They last for more than 4 hours.

I've changed their lifespan to 1 hour so it is more than enough. The brightness has been increased by multiplying the original value by 1.5, more than this is not practical since light may clip through walls a lot more than before.
MPGlowLifetimeMultiplier has been changed to 1.0 so Lighstick duration on MP doesn't get affected.