Improved Lightsticks Vol. I
An attempt to improve the Lightsticks efficiency.
The Lifespan has been increased to 1 hour (in real life they last for more than 4 hours but is not necessary to have that in the game.)
Brightness has been increased to portray how they work IRL, i've always felt that lightsticks in-game lacked some brightness.