Random SwatOfficer delays before entry
This makes the Initiate action little bit random so Red and Blue squads are less perfectly synced. 
For example if they both throw a bang inside a room the bangs are exploding the exact same time... which is odd... with this they are exploding with little delay ... it feels more natural!